[ADMINAPI-1023] Admin Api 2.1.0 and 2.2.0 Unable to create applications with big int ids. (#121)
Update EducationOrganizationId from int to long

+++ Application/EdFi.Ods.AdminApi/Features/Applications/AddApplication.cs
@@ -92,7 +92,7 @@ public class AddApplication : IFeature
         public IEnumerable<int>? ProfileIds { get; set; }
 
         [SwaggerSchema(Description = FeatureConstants.EducationOrganizationIdsDescription, Nullable = false)]
-        public IEnumerable<int>? EducationOrganizationIds { get; set; }
+        public IEnumerable<long>? EducationOrganizationIds { get; set; }
 
         [SwaggerSchema(Description = FeatureConstants.OdsInstanceIdsDescription, Nullable = false)]
         public IEnumerable<int>? OdsInstanceIds { get; set; }
